fix(core): fix generated declarations
